Do all tasks - separate ChatButtonQuery ChatGameQuery InlineButtonQuery and InlineGameQuery from CallbackQuery class - add private wrapQuery to handle both inline and chat queries - some changes to some methods and some fixes ...
