+++ src/EventHandler/InlineGameQuery.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace danog\MadelineProto\EventHandler;
+
+final class InlineGameQuery extends AbstractGameQuery
+{
+}
