+++ src/MTProtoTools/UpdateHandler.php
@@ -27,7 +27,11 @@ use Amp\Http\Client\Response;
 use Amp\TimeoutException;
 use danog\MadelineProto\API;
 use danog\MadelineProto\EventHandler\AbstractMessage;
-use danog\MadelineProto\EventHandler\CallbackQuery;
+use danog\MadelineProto\EventHandler\AbstractQuery;
+use danog\MadelineProto\EventHandler\ChatButtonQuery;
+use danog\MadelineProto\EventHandler\ChatGameQuery;
+use danog\MadelineProto\EventHandler\InlineButtonQuery;
+use danog\MadelineProto\EventHandler\InlineGameQuery;
 use danog\MadelineProto\EventHandler\Message;
 use danog\MadelineProto\EventHandler\Message\ChannelMessage;
 use danog\MadelineProto\EventHandler\Message\GroupMessage;
@@ -335,9 +339,29 @@ trait UpdateHandler
     {
         return match ($update['_']) {
             'updateNewChannelMessage', 'updateNewMessage', 'updateNewScheduledMessage', 'updateEditMessage', 'updateEditChannelMessage' => $this->wrapMessage($update['message']),
-            'updateBotCallbackQuery' => new CallbackQuery($this, $update),
+            'updateBotCallbackQuery','updateInlineBotCallbackQuery' => $this->wrapQuery($update),
             default => null
         };
+    }
+
+    /**
+     * Wrap a Message constructor into an abstract Query object.
+     */
+    private function wrapQuery(array $callback): ?AbstractQuery
+    {
+        if ($callback['_'] == 'updateBotCallbackQuery') {
+            if (!isset($callback['game_short_name']) && !empty($callback['game_short_name'])) {
+                return new ChatGameQuery($this, $callback);
+            }
+            return new ChatButtonQuery($this, $callback);
+        }
+        if ($callback['_'] == 'updateInlineBotCallbackQuery') {
+            if (!isset($callback['game_short_name']) && !empty($callback['game_short_name'])) {
+                return new InlineGameQuery($this, $callback);
+            }
+            return new InlineButtonQuery($this, $callback);
+        }
+        return null;
     }
     /**
      * Wrap a Message constructor into an abstract Message object.
